Check first for existence of favicon
If you don't do this, you'll get an error if the code is run on a page without a favicon in the HTML. Remember that you don't actually have to include a favicon in your HTML, the browser will automatically pull in favicon.ico in the root of the domain, by default. So you should always check before doing the removeChild action.